requirements.txt
pygame-ce==2.5.0
numpy==2.0.0
zengl==2.5.5
pypresence==4.3.0
nest-asyncio==1.6.0
pyinstaller==5.13.0
pygbag==0.9.2